tests/CharacterData-appendData.html

<!DOCTYPE html>
<title>CharacterData.appendData</title>
<link rel="author" title="annevk" href="mailto:annevk@opera.com">
<link rel="help" href="../web-dom-core#dom-CharacterData-appendData">
<script src="testharness.js"></script>
<div id="log"></div>
<script>
function testNode(node) {
  test(function() {
    assert_equals(node.data, "test")
    node.appendData("test")
    assert_equals(node.data, "testtest")
  })
}
test(function() {
  testNode(document.createTextNode("test"))
  testNode(document.createComment("test"))
})
</script>
